Move ResourceResult to results
